+++ R/fix-family.R
@@ -44,12 +44,40 @@
         )
       }
     }
-    tw_pars <- get_tw_ab(fam)
+    tw_pars <- get_tw_params(fam)
     fam$rd <- rd_twlss(a = tw_pars[1], b = tw_pars[2])
   }
 
+  ft <- family_type(fam)
+
+  rd_fun <- switch(
+    EXPR = ft,
+    "cnorm" = rd_gaussian,
+    "cpois" = rd_poisson,
+    "clog"  = rd_logistic,
+    NULL
+  )
+
+  # add the RD fun to the family
+  fam$rd <- rd_fun
+
   # return modified family
   fam
+}
+
+#' @importFrom stats rnorm
+rd_gaussian <- function(mu, wt, scale) {
+  rnorm(length(mu), mean = mu, sd = sqrt(scale / wt))
+}
+
+#' @importFrom stats rpois
+rd_poisson <- function(mu, wt, scale) {
+  rpois(length(mu), lambda = mu)
+}
+
+#' @importFrom stats rlogis
+rd_logistic <- function(mu, wt, scale) {
+  rlogis(length(mu), location = mu, scale = scale)
 }
 
 `fix_family_cdf` <- function(family) {
@@ -71,27 +99,27 @@
   }
 
   # choose a CDF functions
-  qfun <- switch(
+  cdf_fun <- switch(
     EXPR = ft,
     "poisson"  = cdf_poisson,
     "gaussian" = cdf_gaussian,
+    "gaulss"   = cdf_gaulss,
+    "gevlss"   = cdf_gevlss,
+    "gumbls"   = cdf_gumbls,
+    "gammals"  = cdf_gammals,
     "binomial" = cdf_binomial,
     "gamma"    = cdf_gamma,
+    "ziplss"    = cdf_ziplss,
     "scaled_t" = make_cdf_scat(nu = theta[1], sigma = theta[2]),
     "negative_binomial" = make_cdf_nb(theta),
     ## "inverse_gaussian"  = cdf_invgaussian, # FIXME: not sure I trust this yet
-    "beta_regression"   = make_cdf_beta(theta),
-    "tweedie"  = make_cdf_tw(theta, ab = get_tw_ab(family)),
+    "beta_regression"   = make_cdf_beta(theta, eps = betar_eps(family)),
+    "tweedie"  = make_cdf_tw(theta, ab = get_tw_params(family)),
     NULL
   )
 
   # add the CDF fun to the family
-  family$cdf <- qfun
-
-  # don't think we need to throw and error - fix.family.rd doesn't
-  #if (is.null(family$cdf)) {
-  #  stop("No CDF functions available for this family")
-  #}
+  family$cdf <- cdf_fun
 
   # return
   family
@@ -112,6 +140,7 @@
   pbinom(
     q * (wt + as.numeric(wt == 0)), size = wt, prob = mu, log.p = log_p
   )
+  #pbinom(floor(wt * q), wt, mu)
 }
 
 #' @importFrom stats pgamma
@@ -145,6 +174,121 @@
   pinvgaussian(q, mu = mu, wt = wt, scale = scale, log_p = log_p)
 }
 
+#' @importFrom stats pnorm
+`cdf_gaulss` <- function(q, mu, wt, scale, log_p = FALSE) {
+  pnorm(
+    q, mean = mu[, 1, drop = TRUE],
+    sd = 1 / mu[, 2, drop = TRUE], log.p = log_p
+  )
+}
+
+`cdf_gumbls` <- function(q, mu, wt, scale, log_p = FALSE) {
+  gamma <- 0.577215664901533 # euler's constant
+  # mu[, 1] is mean, mu[, 2] is log(beta)
+  # F(x) = exp(-exp(-gamma - \frac{x-mu}{beta}))
+  exp(-exp(-gamma - ((q - mu[, 1]) / exp(mu[ ,2]))))
+}
+
+#' @importFrom stats pgamma
+`cdf_gammals` <- function(q, mu, wt, scale, log_p = FALSE) {
+  # mu[,1] is the mean, mu[,2] is log scale (dispersion)
+  # exp(mu[,2]) == phi
+  phi <- exp(mu[, 2])
+  pgamma(q, shape = 1 / phi, scale = mu[, 1] * phi)
+}
+
+`cdf_gevlss` <- function(q, mu, wt, scale, log_p = FALSE, tol = 1e-6) {
+  # check sigma --- doubt needed as this is deep in mgcv
+  # if (any(mu[, 2] <= 0)) {
+  #   stop("'sigma' must be positive")
+  # }
+  
+  # standardized variable
+  z <- (q - mu[, 1]) / mu[, 2] # (x - mu) / sigma
+  
+  # Allocate result
+  out <- numeric(nrow(mu))
+  
+  # any small xi
+  small <- abs(mu[, 3]) < tol
+  
+  # large xi
+  if (any(!small)) {
+    log_term <- log1p(mu[!small, 3] * z[!small])   # log(1 + xi*z)
+    a <- -log_term / mu[!small, 3]
+    
+    out[!small] <- exp(-exp(a))
+  }
+  
+  # small xi Taylor expansion around 0 for the approaching Gumble case
+  if (any(small)) {
+    a <- -z[small] + 0.5 * mu[small, 3] * z[small]^2
+    out[small] <- exp(-exp(a))
+  }
+  
+  # handle support (important for xi < 0)
+  valid <- 1 + mu[, 3] * z > 0
+  out[!valid & mu[, 3] > 0] <- 0
+  out[!valid & mu[, 3] < 0] <- 1
+  
+  out
+}
+
+# implementation needs checking; ChatGPT used to derive the math for the CDF
+#' @importFrom stats ppois
+`cdf_ziplss` <- function(q, mu, wt, scale, log_p = FALSE) {
+  lambda <- exp(mu[, 1])
+  p <- -expm1(-exp(mu[, 2])) # = 1 - exp(-exp(eta))
+
+  log_cdf <- numeric(nrow(mu))
+  
+  # y < 0
+  log_cdf[q < 0] <- -Inf
+
+  # 0 <= y < 1
+  idx0 <- (q >= 0 & q < 1)
+  log_cdf[idx0] <- log1p(-p[idx0]) # log(1 - p)
+
+  # y >= 1
+  idx <- (q >= 1)
+  
+  if (any(idx)) {
+    y_i <- floor(q[idx])
+    lam <- lambda[idx]
+    p_i <- p[idx]
+    
+    # log P(Y=0)
+    log_p0 <- -lam
+    
+    # log Poisson CDF
+    log_cdfy <- ppois(y_i, lam, log.p = TRUE)
+    
+    # compute log(Fy - p0) safely
+    # log(exp(log_Fy) - exp(log_p0))
+    log_num <- log_cdfy + log1p(-exp(log_p0 - log_cdfy))
+    
+    # log(1 - p0)
+    log_denom <- log1p(-exp(log_p0))
+    
+    # log truncated CDF
+    log_trunc <- log_num - log_denom
+    
+    # combine: log((1-p) + p * trunc)
+    # Use log-sum-exp
+    a <- log1p(-p_i)        # log(1 - p)
+    b <- log(p_i) + log_trunc
+    
+    m <- pmax(a, b)
+    log_cdf[idx] <- m + log(exp(a - m) + exp(b - m))
+  }
+
+  if (isFALSE(log_p)) {
+    log_cdf = exp(log_cdf) # stupid name, I know
+  }
+
+  log_cdf # may not be log_cdf but cdf, and likely won't be because log_p is FALSE
+}
+
 ## For distributions that estimate other parameters, we need to bind the values
 ## of those parameters inside the CDF function in the family. `theta` are then
 ## passed to the `make_cdf_xxx()` function as separate parameters with real
@@ -156,7 +300,7 @@
     # what to do with weights wt? I think this should be (q - mu) / sigma / wt
     # q <- (q - mu) / sigma
     pt(
-      (q - mu) / sigma, # scale the data such that q ~ t_{nu}()
+      (q - mu) / scale, # scale the data such that q ~ t_{nu}()
       df = nu,
       lower.tail = TRUE, log.p = log_p
     )
@@ -176,33 +320,207 @@
 }
 
 #' @importFrom stats pbeta
-`make_cdf_beta` <- function(phi) {
+`make_cdf_beta` <- function(phi, eps) {
   function(q, mu, wt, scale, log_p = FALSE) {
-    pbeta(
+    p <- pbeta(
       q,
       shape1 = phi * mu,
       shape2 = phi * (1 - mu),
       lower.tail = TRUE, log.p = log_p
     )
+    p
   }
 }
 
 #' @importFrom tweedie ptweedie
 `make_cdf_tw` <- function(theta, ab) {
-  function(q, mu, wt, scale, log_p = FALSE) {
-    a <- ab[1] # tweedie lower and upper bounds used in fitting
-    b <- ab[2]
-    # compute tweedie power parameter xi
-    xi <- if (theta > 0) {
-      (b + a * exp(-theta)) / (1 + exp(-theta))
-    } else {
-      (b * exp(theta) + a) / (exp(theta) + 1)
+  fun <- if (length(ab) == 1L) {
+    # here ab is the tweedie power specified in Tweedie()
+    function(p, mu, wt, scale, log_p = FALSE){
+      tweedie::ptweedie(
+        p,
+        mu = mu,
+        phi = scale,
+        xi = ab
+      )
     }
-    tweedie::ptweedie(
-      q,
-      mu = mu,
-      phi = scale, # think to handle weights we need scale / wt
-      xi = xi
-    )
+  } else {
+    function(q, mu, wt, scale, log_p = FALSE) {
+      a <- ab[1] # tweedie lower and upper bounds used in fitting
+      b <- ab[2]
+      # compute tweedie power parameter xi
+      xi <- if (theta > 0) {
+        (b + a * exp(-theta)) / (1 + exp(-theta))
+      } else {
+        (b * exp(theta) + a) / (exp(theta) + 1)
+      }
+      tweedie::ptweedie(
+        q,
+        mu = mu,
+        phi = scale, # think to handle weights we need scale / wt
+        xi = xi
+      )
+    }
   }
+  fun
+}
+
+# only really need this for QQ plots
+#' @importFrom mgcv fix.family.qf
+`fix_family_qf` <- function(family) {
+
+  # try obvious thing first and see if mgcv::fix.family.qf() already handles
+  # family
+  family <- mgcv::fix.family.qf(family)
+
+  # if `family` contains a NULL qf we move on, if it is non-null return early
+  # as it doesn't need fixing
+  if (!is.null(family$qf)) {
+    return(family)
+  }
+
+  # handle special cases - mgcv uses theta as a catchall vector of extra params
+  # for some families. If a family has one or more of these they need to be
+  # embedded in the qf function.
+  # Here we grab theta and decide if it needs transforming
+  ft <- family_type(family)
+  theta <- NULL
+  if (has_theta(family)) {
+    transf <- TRUE
+    if (ft %in% c("tweedie")) { # which don't need transform?
+      transf <- FALSE
+    }
+    theta <- theta(family, transform = transf)
+  }
+
+  # choose QF functions - this is where we explicitly handle the special
+  # families; they need a make_qf_foo function, while normal families
+  # with no extra parameters just need a qf_foo function
+  qfun <- switch(
+    EXPR = ft,
+    "scaled_t" = make_qf_scat(nu = theta[1], sigma = theta[2]),
+    "tweedie"  = make_qf_tw(theta, ab = get_tw_params(family)),
+    #"nb"       = make_qf_betar(phi = theta),
+    #"betar"    = make_qf_betar(theta),
+    "gaulss"   = qf_gaulss,
+    "gevlss"   = qf_gevlss,
+    "gumbls"   = qf_gumbls,
+    "gammals"  = qf_gammals,
+    "ziplss"   = qf_ziplss,
+    NULL # if don't handle family, return NULL as qfun so family unchanged
+  )
+
+  # add the QF fun to the family
+  family$qf <- qfun
+
+  # return
+  family
+}
+
+#' @importFrom stats qnorm
+`qf_gaulss` <- function(p, mu, wt, scale, log_p = FALSE) {
+  qnorm(
+    p,
+    mean = mu[, 1, drop = TRUE],
+    sd = 1 / mu[, 2, drop = TRUE],
+    log.p = log_p
+  )
+}
+
+`qf_gevlss` <- function(p, mu, wt, log_p = FALSE, tol = 1e-6) {
+  # checks, but I doubt these are needed as this is deep in mgcv
+  #if (any(p <= 0 | p >= 1)) {
+  #  stop("All probabilities 'p' must be in (0, 1)")
+  #}
+  #if (any(mu[, 2] <= 0)) {
+  #  stop("All 'sigma' must be positive")
+  #}
+
+  # extract vectors --- perhaps not if data is big?
+  # sigma <- mu[, 2]
+  # xi    <- mu[, 3]
+  # mu    <- mu[, 1]
+
+  t <- -log(p)
+  logt <- log(t)
+
+  # |xi| well away from zero so use textbook formula
+  small <- abs(mu[, 3]) < tol
+  # large <- !small
+
+  out <- numeric(length(p))
+
+  # textbook formula
+  if (any(!small)) {
+    out[!small] <- mu[!small, 1] + mu[!small, 2] *
+      expm1(-mu[!small, 3] * logt[!small]) / mu[!small, 3]
+  }
+
+  # use Taylor expansion around xi = 0
+  if (any(small)) {
+    out[small] <- mu[small, 1] + mu[small, 2] * 
+      (-logt[small] + 0.5 * mu[small, 3] * logt[small]^2)
+  }
+
+  out
+}
+
+# mgcv uses a mean-centred parameterisation
+qf_gumbls <- function(p, mu, wt, scale, log_p = FALSE) {
+  gamma <- 0.577215664901533 # euler's constant
+  # mu[, 1] is mean, mu[, 2] is log(beta)
+  mu[, 1] - exp(mu[, 2]) * (gamma + log(-log(p)))
+}
+
+#' @importFrom stats qgamma
+`qf_gammals` <- function(p, mu, wt, scale, log_p = FALSE) {
+  # mu[,1] is the mean, mu[,2] is log scale (dispersion)
+  # exp(mu[,2]) == phi
+  phi <- exp(mu[, 2])
+  qgamma(p, shape = 1 / phi, scale = mu[, 1] * phi)
+}
+
+#' @importFrom stats pt
+`make_qf_scat` <- function(nu, sigma) {
+  function(p, mu, wt, scale, log_p = FALSE) {
+    qt(
+      p,
+      df = nu,
+      lower.tail = TRUE,
+      log.p = log_p
+    ) * scale + mu
+  }
+}
+
+#' @importFrom tweedie qtweedie
+`make_qf_tw` <- function(theta, ab) {
+  fun <- if (length(ab) == 1L) {
+    # here ab is the tweedie power specified in Tweedie()
+    function(p, mu, wt, scale, log_p = FALSE){
+      tweedie::qtweedie(
+        p,
+        mu = mu,
+        phi = scale,
+        xi = ab
+      )
+    }
+  } else {
+    function(p, mu, wt, scale, log_p = FALSE) {
+      a <- ab[1] # tweedie lower and upper bounds used in fitting
+      b <- ab[2]
+      # compute tweedie power parameter xi
+      xi <- if (theta > 0) {
+        (b + a * exp(-theta)) / (1 + exp(-theta))
+      } else {
+        (b * exp(theta) + a) / (exp(theta) + 1)
+      }
+      tweedie::qtweedie(
+        p,
+        mu = mu,
+        phi = scale, # think to handle weights we need scale / wt
+        xi = xi
+      )
+    }
+  }
+  fun
 }